requirements_siglip.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
tokenizers>=0.15.0
datasets>=2.14.0
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
wandb>=0.16.0
tqdm>=4.65.0
accelerate>=0.24.0
sentencepiece>=0.1.99
protobuf>=4.21.0 